+++ CHANGELOG.markdown
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Redesigned tree / collapsing header API
+
 ## [0.3.0] - 2020-02-15
 
 ### Added